rdrf #1887 method decorator added
