+++ rdrf/rdrf/helpers/cache_utils.py
@@ -1,0 +1,51 @@
+import logging
+import re
+from django.core.cache import caches
+
+logger = logging.getLogger(__name__)
+
+illegal_characters = r"[:\"{}' ]"
+
+
+def use_cache(function):
+    def wrapper(*args, **kwargs):
+        key = f"{function.__name__}_{re.sub(illegal_characters, '', str(kwargs))}"
+        query_cache = caches["queries"]
+        if key in query_cache:
+            return query_cache.get(key)
+        else:
+            result = function(*args, **kwargs)
+            query_cache.set(key, result)
+            return result
+    return wrapper
+
+
+def use_object_cache(method):
+    def wrapper(*args, **kwargs):
+        self = args[0]
+        obj = re.sub(illegal_characters, '', str(self))
+        kw = re.sub(illegal_characters, '', str(kwargs))
+        key = f"{method.__qualname__}_{obj}_id{str(self.pk)}_{kw}"
+        query_cache = caches["queries"]
+        if key in query_cache:
+            logger.info(f"Fetching from Cache: {key}")
+            return query_cache.get(key)
+        else:
+            result = method(*args, **kwargs)
+            logger.info(f"Caching: {key}")
+            query_cache.set(key, result)
+            return result
+    return wrapper
+
+
+def get_cache(key=None):
+    query_cache = caches["queries"]
+    if key in query_cache:
+        return query_cache.get(key)
+    else:
+        return None
+
+
+def set_cache(key=None, value=None):
+    query_cache = caches["queries"]
+    query_cache.set(key, value)
